requirements.txt
llama-cpp-python==0.2.18
gradio==4.4.1